Init for V2

+++ Client/src/main/java/org/ow2/frascati/examples/helloworld/annotated/GenerarPts.java
@@ -32,5 +32,5 @@ import org.osoa.sca.annotations.Service;
 @Service
 public interface GenerarPts
 {
-    long generarPts(long seed, long ptsTotales);
+    int generarPts(long seed, long ptsTotales);
 }
